refactor: extract TemplateParser class from TemplateElement
- Move all parsing logic (resolve* methods, TemplateResolutionContext,
  deprecated event syntax tracking) into a standalone TemplateParser class
  in template-parser.ts
- TemplateElement now focuses on custom element lifecycle: registration,
  options, callbacks, ObserverMap/AttributeMap wiring, template assignment
- TemplateParser exposes parse() and createTemplate() public methods,
  making it independently testable without DOM and reusable outside
  <f-template>
- Export TemplateParser from package index
- Update DESIGN.md with new architecture, sequence diagram, and method
  decomposition tables
