+++ packages/fast-html/src/components/template.ts
@@ -1,64 +1,19 @@
 import {
     attr,
-    children,
     ElementController,
-    elements,
     FAST,
     FASTElement,
     FASTElementDefinition,
     fastElementRegistry,
-    ref,
-    repeat,
-    slotted,
     type TemplateLifecycleCallbacks,
-    ViewTemplate,
-    when,
 } from "@microsoft/fast-element";
 import "@microsoft/fast-element/install-hydratable-view-templates.js";
 import { Message } from "../interfaces.js";
 import { AttributeMap } from "./attribute-map.js";
 import { ObserverMap } from "./observer-map.js";
 import { Schema } from "./schema.js";
-import {
-    type AttributeDirective,
-    type AttributeDirectiveBindingBehaviorConfig,
-    bindingResolver,
-    type ChainedExpression,
-    contextPrefixDot,
-    type DataBindingBehaviorConfig,
-    eventArgAccessor,
-    getBooleanBinding,
-    getExpressionChain,
-    getNextBehavior,
-    getRootPropertyName,
-    parseEventArgs,
-    type TemplateDirectiveBehaviorConfig,
-    transformInnerHTML,
-} from "./utilities.js";
-
-/**
- * The return type for {@link TemplateElement.resolveStringsAndValues}.
- */
-interface ResolvedStringsAndValues {
-    strings: Array<string>;
-    values: Array<any>;
-}
-
-/**
- * Encapsulates the stable context fields that flow through the recursive
- * template resolution pipeline. Keeps method signatures lean and makes it
- * easy to add new context without touching every call site.
- *
- * `rootPropertyName` is intentionally **excluded** because it is
- * selectively mutated per branch and must not leak across siblings.
- */
-interface TemplateResolutionContext {
-    self: boolean;
-    parentContext: string | null;
-    level: number;
-    schema: Schema;
-    observerMap?: ObserverMap;
-}
+import { TemplateParser } from "./template-parser.js";
+import { eventArgAccessor, transformInnerHTML } from "./utilities.js";
 
 /**
  * Values for the observerMap element option.
@@ -218,7 +173,11 @@ export interface HydrationLifecycleCallbacks extends TemplateLifecycleCallbacks 
 }
 
 /**
- * The <f-template> custom element that will provide view logic to the element
+ * The <f-template> custom element that will provide view logic to the element.
+ *
+ * Acts as the bridge between declarative HTML templates and the FAST element
+ * registry. Lifecycle orchestration (registration, options, callbacks) lives
+ * here; template parsing is delegated to {@link TemplateParser}.
  */
 class TemplateElement extends FASTElement {
     /**
@@ -251,14 +210,6 @@ class TemplateElement extends FASTElement {
      * Metadata containing JSON schema for properties on a custom element
      */
     private schema?: Schema;
-
-    /**
-     * Whether the template contains deprecated "e" event argument usage.
-     * Set during template processing; checked after evaluation to emit a
-     * single warning per template.
-     */
-    // TODO: remove per https://github.com/microsoft/fast/issues/7314
-    private _hasDeprecatedEventSyntax = false;
 
     /**
      * Lifecycle callbacks for hydration events
@@ -385,21 +336,15 @@ class TemplateElement extends FASTElement {
                 TemplateElement.lifecycleCallbacks.templateWillUpdate?.(name);
 
                 const innerHTML = transformInnerHTML(this.innerHTML);
+                const parser = new TemplateParser();
 
-                // Cache paths during template processing (pass undefined if observerMap is not available)
-                const { strings, values } = await this.resolveStringsAndValues(
-                    null,
+                const { strings, values } = await parser.parse(
                     innerHTML,
-                    {
-                        self: false,
-                        parentContext: null,
-                        level: 0,
-                        schema: this.schema as Schema,
-                        observerMap: this.observerMap,
-                    },
+                    this.schema as Schema,
+                    this.observerMap,
                 );
 
-                if (this._hasDeprecatedEventSyntax) {
+                if (parser.hasDeprecatedEventSyntax) {
                     console.warn(
                         `[fast-html] Using "e" as an event argument is deprecated` +
                             ` in component "${name}".` +
@@ -425,7 +370,7 @@ class TemplateElement extends FASTElement {
 
                     // All new elements will get the updated template
                     // This assignment triggers the Observable notification → callbacks fire
-                    registeredFastElement.template = this.resolveTemplateOrBehavior(
+                    registeredFastElement.template = parser.createTemplate(
                         strings,
                         values,
                     );
@@ -438,572 +383,6 @@ class TemplateElement extends FASTElement {
                 throw FAST.error(Message.noTemplateProvided, { name: this.name });
             }
         });
-    }
-
-    /**
-     * Resolve strings and values from an innerHTML string
-     * @param rootPropertyName - The root property name for schema registration.
-     * @param innerHTML - The innerHTML.
-     * @param context - The template resolution context.
-     */
-    private async resolveStringsAndValues(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        context: TemplateResolutionContext,
-    ): Promise<ResolvedStringsAndValues> {
-        const strings: any[] = [];
-        const values: any[] = []; // these can be bindings, directives, etc.
-        await this.resolveInnerHTML(
-            rootPropertyName,
-            innerHTML,
-            strings,
-            values,
-            context,
-        );
-
-        (strings as any).raw = strings.map(value => String.raw({ raw: value }));
-
-        return {
-            strings,
-            values,
-        };
-    }
-
-    /**
-     * Resolve a template or behavior
-     * @param strings - The strings array.
-     * @param values - The interpreted values.
-     */
-    private resolveTemplateOrBehavior(
-        strings: Array<string>,
-        values: Array<any>,
-    ): ViewTemplate<any, any> {
-        return ViewTemplate.create(strings, values);
-    }
-
-    /**
-     * Resolve a template directive
-     * @param behaviorConfig - The directive behavior configuration object.
-     * @param externalValues - The interpreted values from the parent.
-     * @param innerHTML - The innerHTML.
-     * @param self - Indicates that this should refer to itself instead of a property when creating bindings.
-    /**
-     * Resolve a template directive (when/repeat).
-     * @param rootPropertyName - The root property name for schema registration.
-     * @param behaviorConfig - The directive behavior configuration object.
-     * @param externalValues - The interpreted values from the parent.
-     * @param innerHTML - The innerHTML.
-     * @param context - The template resolution context.
-     */
-    private async resolveTemplateDirective(
-        rootPropertyName: string | null,
-        behaviorConfig: TemplateDirectiveBehaviorConfig,
-        externalValues: Array<any>,
-        innerHTML: string,
-        context: TemplateResolutionContext,
-    ): Promise<void> {
-        switch (behaviorConfig.name) {
-            case "when": {
-                const expressionChain = getExpressionChain(behaviorConfig.value);
-
-                const whenLogic = getBooleanBinding(
-                    rootPropertyName,
-                    expressionChain as ChainedExpression,
-                    context.parentContext,
-                    context.level,
-                    context.schema,
-                );
-
-                const { strings, values } = await this.resolveStringsAndValues(
-                    rootPropertyName,
-                    innerHTML.slice(
-                        behaviorConfig.openingTagEndIndex,
-                        behaviorConfig.closingTagStartIndex,
-                    ),
-                    context,
-                );
-
-                externalValues.push(
-                    when(whenLogic, this.resolveTemplateOrBehavior(strings, values)),
-                );
-
-                break;
-            }
-            case "repeat": {
-                const valueAttr = behaviorConfig.value.split(" "); // syntax {{x in y}}
-                const updatedLevel = context.level + 1;
-
-                rootPropertyName = getRootPropertyName(
-                    rootPropertyName,
-                    valueAttr[2],
-                    context.parentContext,
-                    behaviorConfig.name,
-                );
-                const binding = bindingResolver(
-                    null,
-                    rootPropertyName,
-                    valueAttr[2],
-                    context.parentContext,
-                    behaviorConfig.name,
-                    context.schema,
-                    valueAttr[0],
-                    context.level,
-                );
-
-                const repeatContext: TemplateResolutionContext = {
-                    self: true,
-                    parentContext: valueAttr[0],
-                    level: updatedLevel,
-                    schema: context.schema,
-                    observerMap: context.observerMap,
-                };
-
-                const { strings, values } = await this.resolveStringsAndValues(
-                    rootPropertyName,
-                    innerHTML.slice(
-                        behaviorConfig.openingTagEndIndex,
-                        behaviorConfig.closingTagStartIndex,
-                    ),
-                    repeatContext,
-                );
-
-                externalValues.push(
-                    repeat(
-                        (x, c) => binding(x, c),
-                        this.resolveTemplateOrBehavior(strings, values),
-                    ),
-                );
-
-                break;
-            }
-        }
-    }
-
-    /**
-     * Resolve a template directive
-     * @param name - The name of the directive.
-     * @param propName - The property name to pass to the directive.
-     * @param externalValues - The interpreted values from the parent.
-     */
-    private async resolveAttributeDirective(
-        name: AttributeDirective,
-        propName: string,
-        externalValues: Array<any>,
-    ) {
-        switch (name) {
-            case "children": {
-                externalValues.push(children(propName));
-
-                break;
-            }
-            case "slotted": {
-                const parts = propName.trim().split(" filter ");
-                const slottedOption = {
-                    property: parts[0],
-                };
-
-                if (parts[1]) {
-                    if (parts[1].startsWith("elements(")) {
-                        let params = parts[1].replace("elements(", "");
-                        params = params.substring(0, params.lastIndexOf(")"));
-                        Object.assign(slottedOption, {
-                            filter: elements(params || undefined),
-                        });
-                    }
-                }
-
-                externalValues.push(slotted(slottedOption));
-
-                break;
-            }
-            case "ref": {
-                externalValues.push(ref(propName));
-
-                break;
-            }
-        }
-    }
-
-    /**
-     * Resolve an access binding — shared by content bindings, boolean-attribute
-     * fallback, and default attribute bindings.
-     * @returns An object with the resolved binding function and the updated rootPropertyName.
-     */
-    private resolveAccessBinding(
-        rootPropertyName: string | null,
-        propName: string,
-        previousStrings: string,
-        context: TemplateResolutionContext,
-    ): {
-        binding: (x: any, c: any) => any;
-        rootPropertyName: string | null;
-    } {
-        rootPropertyName = getRootPropertyName(
-            rootPropertyName,
-            propName,
-            context.parentContext,
-            "access",
-        );
-        const resolved = bindingResolver(
-            previousStrings,
-            rootPropertyName,
-            propName,
-            context.parentContext,
-            "access",
-            context.schema,
-            context.parentContext,
-            context.level,
-        );
-        return {
-            binding: (x: any, c: any) => resolved(x, c),
-            rootPropertyName,
-        };
-    }
-
-    /**
-     * Resolve an event binding (the "@" aspect).
-     * @returns An object with the event binding function and the updated rootPropertyName.
-     */
-    private resolveEventBinding(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        behaviorConfig: DataBindingBehaviorConfig,
-        strings: Array<string>,
-        context: TemplateResolutionContext,
-    ): {
-        binding: (x: any, c: any) => any;
-        rootPropertyName: string | null;
-    } {
-        const bindingHTML = innerHTML.slice(
-            behaviorConfig.openingEndIndex,
-            behaviorConfig.closingStartIndex,
-        );
-        const openingParenthesis = bindingHTML.indexOf("(");
-        const closingParenthesis = bindingHTML.indexOf(")");
-        const propName = innerHTML.slice(
-            behaviorConfig.openingEndIndex,
-            behaviorConfig.closingStartIndex -
-                (closingParenthesis - openingParenthesis) -
-                1,
-        );
-        const type = "event";
-        rootPropertyName = getRootPropertyName(
-            rootPropertyName,
-            propName,
-            context.parentContext,
-            type,
-        );
-        const argsString = bindingHTML.slice(openingParenthesis + 1, closingParenthesis);
-        const resolved = bindingResolver(
-            strings.join(""),
-            rootPropertyName,
-            propName,
-            context.parentContext,
-            type,
-            context.schema,
-            context.parentContext,
-            context.level,
-        );
-        const isContextPath = propName.startsWith(contextPrefixDot);
-        const getOwner = isContextPath
-            ? (_x: any, c: any) => {
-                  const ownerPath = propName.split(".").slice(1, -1);
-                  return ownerPath.reduce((prev: any, item: string) => prev?.[item], c);
-              }
-            : (x: any, _c: any) => x;
-
-        const parsedArgs = parseEventArgs(argsString);
-
-        if (parsedArgs.some(a => a.type === "deprecated-event")) {
-            this._hasDeprecatedEventSyntax = true;
-        }
-
-        const argResolvers = parsedArgs.map((parsedArg): ((x: any, c: any) => any) => {
-            switch (parsedArg.type) {
-                case "event":
-                case "deprecated-event":
-                    return (_x, c) => c.event;
-                case "context":
-                    return (_x, c) => c;
-                case "binding":
-                    return bindingResolver(
-                        strings.join(""),
-                        rootPropertyName,
-                        parsedArg.rawArg!,
-                        context.parentContext,
-                        type,
-                        context.schema,
-                        context.parentContext,
-                        context.level,
-                    );
-            }
-        });
-
-        return {
-            binding: (x: any, c: any) =>
-                resolved(x, c).bind(getOwner(x, c))(
-                    ...argResolvers.map(resolve => resolve(x, c)),
-                ),
-            rootPropertyName,
-        };
-    }
-
-    /**
-     * Resolve a content data binding (`{{expression}}` in text content).
-     */
-    private async resolveContentBinding(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        strings: Array<string>,
-        values: Array<any>,
-        behaviorConfig: DataBindingBehaviorConfig,
-        context: TemplateResolutionContext,
-    ): Promise<void> {
-        strings.push(innerHTML.slice(0, behaviorConfig.openingStartIndex));
-        const propName = innerHTML.slice(
-            behaviorConfig.openingEndIndex,
-            behaviorConfig.closingStartIndex,
-        );
-        const result = this.resolveAccessBinding(
-            rootPropertyName,
-            propName,
-            strings.join(""),
-            context,
-        );
-        rootPropertyName = result.rootPropertyName;
-        values.push(result.binding);
-        await this.resolveInnerHTML(
-            rootPropertyName,
-            innerHTML.slice(behaviorConfig.closingEndIndex, innerHTML.length),
-            strings,
-            values,
-            context,
-        );
-    }
-
-    /**
-     * Resolve an attribute data binding (`{{expression}}` in an HTML attribute).
-     * Dispatches to event, expression, or access binding handlers based on aspect.
-     */
-    private async resolveAttributeBinding(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        strings: Array<string>,
-        values: Array<any>,
-        behaviorConfig: DataBindingBehaviorConfig,
-        context: TemplateResolutionContext,
-    ): Promise<void> {
-        strings.push(innerHTML.slice(0, behaviorConfig.openingStartIndex));
-        let attributeBinding;
-
-        const aspect =
-            behaviorConfig.subtype === "attribute" ? behaviorConfig.aspect : null;
-
-        switch (aspect) {
-            case "@": {
-                const result = this.resolveEventBinding(
-                    rootPropertyName,
-                    innerHTML,
-                    behaviorConfig,
-                    strings,
-                    context,
-                );
-                attributeBinding = result.binding;
-                rootPropertyName = result.rootPropertyName;
-                break;
-            }
-            case "?": {
-                const expression = innerHTML.slice(
-                    behaviorConfig.openingEndIndex,
-                    behaviorConfig.closingStartIndex,
-                );
-                const expressionChain = getExpressionChain(expression);
-
-                if (expressionChain?.expression.operator) {
-                    attributeBinding = getBooleanBinding(
-                        rootPropertyName,
-                        expressionChain as ChainedExpression,
-                        context.parentContext,
-                        context.level,
-                        context.schema,
-                    );
-                } else {
-                    const propName = innerHTML.slice(
-                        behaviorConfig.openingEndIndex,
-                        behaviorConfig.closingStartIndex,
-                    );
-                    const result = this.resolveAccessBinding(
-                        rootPropertyName,
-                        propName,
-                        strings.join(""),
-                        context,
-                    );
-                    attributeBinding = result.binding;
-                    rootPropertyName = result.rootPropertyName;
-                }
-
-                break;
-            }
-            default: {
-                const propName = innerHTML.slice(
-                    behaviorConfig.openingEndIndex,
-                    behaviorConfig.closingStartIndex,
-                );
-                const result = this.resolveAccessBinding(
-                    rootPropertyName,
-                    propName,
-                    strings.join(""),
-                    context,
-                );
-                attributeBinding = result.binding;
-                rootPropertyName = result.rootPropertyName;
-            }
-        }
-
-        values.push(attributeBinding);
-
-        await this.resolveInnerHTML(
-            rootPropertyName,
-            innerHTML.slice(behaviorConfig.closingEndIndex, innerHTML.length),
-            strings,
-            values,
-            context,
-        );
-    }
-
-    /**
-     * Resolve an attribute directive binding (`f-children`, `f-slotted`, `f-ref`).
-     */
-    private async resolveAttributeDirectiveBinding(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        strings: Array<string>,
-        values: Array<any>,
-        behaviorConfig: AttributeDirectiveBindingBehaviorConfig,
-        context: TemplateResolutionContext,
-    ): Promise<void> {
-        strings.push(
-            innerHTML.slice(
-                0,
-                behaviorConfig.openingStartIndex - behaviorConfig.name.length - 4,
-            ),
-        );
-        const propName = innerHTML.slice(
-            behaviorConfig.openingEndIndex,
-            behaviorConfig.closingStartIndex,
-        );
-        await this.resolveAttributeDirective(behaviorConfig.name, propName, values);
-        await this.resolveInnerHTML(
-            rootPropertyName,
-            innerHTML.slice(behaviorConfig.closingEndIndex + 1, innerHTML.length),
-            strings,
-            values,
-            context,
-        );
-    }
-
-    /**
-     * Dispatcher for data binding resolution. Routes to the appropriate handler
-     * based on the binding subtype.
-     */
-    private async resolveDataBinding(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        strings: Array<string>,
-        values: Array<any>,
-        behaviorConfig: DataBindingBehaviorConfig,
-        context: TemplateResolutionContext,
-    ): Promise<void> {
-        switch (behaviorConfig.subtype) {
-            case "content":
-                return this.resolveContentBinding(
-                    rootPropertyName,
-                    innerHTML,
-                    strings,
-                    values,
-                    behaviorConfig,
-                    context,
-                );
-            case "attribute":
-                return this.resolveAttributeBinding(
-                    rootPropertyName,
-                    innerHTML,
-                    strings,
-                    values,
-                    behaviorConfig,
-                    context,
-                );
-            case "attributeDirective":
-                return this.resolveAttributeDirectiveBinding(
-                    rootPropertyName,
-                    innerHTML,
-                    strings,
-                    values,
-                    behaviorConfig,
-                    context,
-                );
-        }
-    }
-
-    /**
-     * Resolver of the innerHTML string. Finds the next binding or directive
-     * in the HTML and dispatches to the appropriate handler.
-     * @param rootPropertyName - The root property name for schema registration.
-     * @param innerHTML - The innerHTML to parse.
-     * @param strings - The strings array (accumulates literal HTML segments).
-     * @param values - The values array (accumulates binding functions and directives).
-     * @param context - The template resolution context.
-     */
-    private async resolveInnerHTML(
-        rootPropertyName: string | null,
-        innerHTML: string,
-        strings: Array<string>,
-        values: Array<any>,
-        context: TemplateResolutionContext,
-    ): Promise<void> {
-        const behaviorConfig = getNextBehavior(innerHTML);
-
-        if (behaviorConfig === null) {
-            strings.push(innerHTML);
-        } else {
-            switch (behaviorConfig.type) {
-                case "dataBinding": {
-                    await this.resolveDataBinding(
-                        rootPropertyName,
-                        innerHTML,
-                        strings,
-                        values,
-                        behaviorConfig,
-                        context,
-                    );
-
-                    break;
-                }
-                case "templateDirective": {
-                    strings.push(innerHTML.slice(0, behaviorConfig.openingTagStartIndex));
-                    await this.resolveTemplateDirective(
-                        rootPropertyName,
-                        behaviorConfig,
-                        values,
-                        innerHTML,
-                        context,
-                    );
-
-                    await this.resolveInnerHTML(
-                        rootPropertyName,
-                        innerHTML.slice(
-                            behaviorConfig.closingTagEndIndex,
-                            innerHTML.length,
-                        ),
-                        strings,
-                        values,
-                        context,
-                    );
-
-                    break;
-                }
-            }
-        }
     }
 }
